[impl] Implement reusable hook
