+++ app/globals.css
@@ -143,7 +143,7 @@
 
 @layer utilities {
   .glass-effect {
-    @apply rounded-3xl backdrop-blur-sm backdrop-saturate-100 shadow-[0_8px_32px_0_rgba(0,0,0,0.6),inset_0_1px_0_0_rgba(255,255,255,0.2)] transition-all duration-500;
+    @apply rounded-3xl backdrop-blur-sm backdrop-saturate-100 shadow-[0_8px_32px_0_#0d2626,inset_0_1px_0_0_rgba(255,255,255,0.2)] transition-all duration-500;
   }
   .container {
     @apply mx-auto max-w-screen-2xl px-4 md:px-6 lg:px-8;
