+++ components/Header.tsx
@@ -1,26 +1,18 @@
 "use client";
 
+import useShowNavItems from "@/hooks/useShowNavItems";
 import Image from "next/image";
 import Link from "next/link";
-import { usePathname } from "next/navigation";
-import { useEffect, useState } from "react";
 import HomeCombobox from "./HomeCombobox";
 import NavItems from "./NavItems";
 import UserDropdown from "./UserDropdown";
 
 const Header = () => {
-  const pathname = usePathname();
-  const [showNavItems, setShowNavItems] = useState(false);
-
-  const isHome = pathname === "/";
-
-  useEffect(() => {
-    setShowNavItems(!isHome);
-  }, [isHome]);
+  const showNavItems = useShowNavItems();
 
   return (
     <header className="sticky top-6 z-50 w-full px-6">
-      <div className="container header-wrapper glass-effect">
+      <div className="container header-wrapper glass-effect md:gap-2">
         <Link href="/">
           <Image
             src="/assets/logo/logo.png"
@@ -30,7 +22,7 @@ const Header = () => {
             className="w-auto cursor-pointer"
           />
         </Link>
-        <nav className="hidden sm:flex items-center gap-4 justify-center">
+        <nav className="hidden sm:flex items-center gap-2 justify-center">
           <HomeCombobox />
           {showNavItems && <NavItems />}
         </nav>
